refactor: UI 로직을 features/loading으로 분리
- features/loading/hooks/useRecommendationWithProgress 생성
- entities/location/hooks/useLocations에서 isProgressLoading, getRecommendationFull 제거
- entities/location/hooks/useLocations에 setData 추가
- IndexPage, MeetingForm에서 useRecommendationWithProgress 사용
- FSD 원칙 준수: entities는 순수 데이터만, UI 로직은 features에
